+++ GUI/Controls/ModInfo/Relationships.Designer.cs
@@ -83,6 +83,7 @@ namespace CKAN.GUI
             this.DependsGraphTree.ImageList.Images.Add("Suggests", global::CKAN.GUI.EmbeddedImages.info);
             this.DependsGraphTree.ImageList.Images.Add("Supports", global::CKAN.GUI.EmbeddedImages.smile);
             this.DependsGraphTree.ImageList.Images.Add("Conflicts", global::CKAN.GUI.EmbeddedImages.alert);
+            this.DependsGraphTree.ImageList.Images.Add("InstalledConflicts", global::CKAN.GUI.EmbeddedImages.stop);
             //
             // LegendTable
             //
@@ -102,14 +103,14 @@ namespace CKAN.GUI
             this.LegendTable.Controls.Add(this.LegendProvidesLabel, 1, 1);
             this.LegendTable.Controls.Add(this.LegendDependsImage, 0, 2);
             this.LegendTable.Controls.Add(this.LegendDependsLabel, 1, 2);
-            this.LegendTable.Controls.Add(this.LegendRecommendsImage, 2, 1);
-            this.LegendTable.Controls.Add(this.LegendRecommendsLabel, 3, 1);
-            this.LegendTable.Controls.Add(this.LegendSuggestsImage, 2, 2);
-            this.LegendTable.Controls.Add(this.LegendSuggestsLabel, 3, 2);
-            this.LegendTable.Controls.Add(this.LegendSupportsImage, 4, 1);
-            this.LegendTable.Controls.Add(this.LegendSupportsLabel, 5, 1);
-            this.LegendTable.Controls.Add(this.LegendConflictsImage, 4, 2);
-            this.LegendTable.Controls.Add(this.LegendConflictsLabel, 5, 2);
+            this.LegendTable.Controls.Add(this.LegendConflictsImage, 2, 1);
+            this.LegendTable.Controls.Add(this.LegendConflictsLabel, 3, 1);
+            this.LegendTable.Controls.Add(this.LegendRecommendsImage, 2, 2);
+            this.LegendTable.Controls.Add(this.LegendRecommendsLabel, 3, 2);
+            this.LegendTable.Controls.Add(this.LegendSuggestsImage, 4, 1);
+            this.LegendTable.Controls.Add(this.LegendSuggestsLabel, 5, 1);
+            this.LegendTable.Controls.Add(this.LegendSupportsImage, 4, 2);
+            this.LegendTable.Controls.Add(this.LegendSupportsLabel, 5, 2);
             this.LegendTable.Dock = System.Windows.Forms.DockStyle.Bottom;
             this.LegendTable.Location = new System.Drawing.Point(0, 0);
             this.LegendTable.Name = "LegendTable";
@@ -132,14 +133,14 @@ namespace CKAN.GUI
             // LegendIncompatibleLabel
             //
             this.LegendIncompatibleLabel.AutoSize = true;
-            this.LegendIncompatibleLabel.ForeColor = System.Drawing.Color.Red;
+            this.LegendIncompatibleLabel.ForeColor = System.Drawing.SystemColors.GrayText;
             this.LegendIncompatibleLabel.Padding = new System.Windows.Forms.Padding(0, 0, 0, 6);
             resources.ApplyResources(this.LegendIncompatibleLabel, "LegendIncompatibleLabel");
             //
             // LegendVirtualLabel
             //
             this.LegendVirtualLabel.AutoSize = true;
-            this.LegendVirtualLabel.ForeColor = System.Drawing.SystemColors.GrayText;
+            this.LegendVirtualLabel.Font = new System.Drawing.Font(System.Drawing.SystemFonts.DefaultFont, System.Drawing.FontStyle.Italic);
             this.LegendVirtualLabel.Padding = new System.Windows.Forms.Padding(0, 0, 0, 6);
             resources.ApplyResources(this.LegendVirtualLabel, "LegendVirtualLabel");
             //